# Recycle size 1 `value` efficiently in `vec_assign()`