… src/slice-assign-array.c
… tests/testthat/test-slice-assign.R