I'm really after the memory improvements here, but the performance improvements are nice too!

``` r
library(vctrs)

x <- seq_len(1e6) + 0L
i <- sample(1e6)
value <- 0L

# Before
bench::mark(vec_assign(x, i, value))
#> # A tibble: 1 × 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_assign(x, i, value)   2.04ms   2.33ms      422.    7.63MB     161.

# After
bench::mark(vec_assign(x, i, value))
#> # A tibble: 1 × 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_assign(x, i, value)   1.69ms   1.87ms      528.    3.82MB     52.6

# Base
bench::mark(`[<-`(x, i, value))
#> # A tibble: 1 × 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 `[<-`(x, i, value)   2.49ms   2.63ms      378.    3.81MB     56.5
```